requirements.txt
# PaddleOCR Production System - All Dependencies
# Last Updated: 2025

# Core PaddleOCR with CPU optimizations
paddlepaddle>=2.6.1
paddleocr>=2.7.0.3

# Image processing
opencv-python>=4.8.1.78
Pillow>=10.1.0
scikit-image>=0.22.0

# Database handling
dbfread>=2.0.7
dbfwrite>=1.0.1
pandas>=2.1.4  # For result analysis

# Utilities
numpy>=1.24.3
tqdm>=4.66.1  # Progress bars
psutil>=5.9.6  # System monitoring
python-dateutil>=2.8.2

# Optional - For distributed processing
redis>=5.0.1  # If using queue system
celery>=5.3.4  # If using distributed workers

# Performance monitoring
prometheus-client>=0.19.0  # Metrics

numpy==1.24.3
opencv-python==4.8.1.78
paddleocr==2.7.0.3
imgaug==0.4.0